move deploy to it's own pkg to avoid errors
