+++ pkg/deploy/deploy_test.go
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package main
+package deploy
 
 import (
 	"testing"
